Move transactions index api into /archive
