+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/module/ApiModule.java
@@ -157,10 +157,6 @@ public class ApiModule extends AbstractModule {
 		}
 		endpointStatus.put("version", versionEnable);
 
-		var transactionIndex = properties.get("api.developer.transaction_index.enable", false);
-		if (transactionIndex) {
-			install(new TransactionIndexApiModule());
-		}
 		var developerEnable = properties.get("api.developer.enable", true);
 		if (developerEnable) {
 			install(new DeveloperEndpointModule());
