+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/transactions/lookup/BerkeleyTransactionsByIdStore.java
@@ -114,7 +114,7 @@ import static com.sleepycat.je.LockMode.DEFAULT;
 import static com.sleepycat.je.OperationStatus.SUCCESS;
 
 public final class BerkeleyTransactionsByIdStore implements BerkeleyAdditionalStore {
-	private static final String TXN_ID_DB_NAME = "radix.transactions_by_id_db";
+	private static final String TXN_ID_DB_NAME = "radix.transactions";
 	private Database txnIdDatabase; // Txns by AID; Append-only
 	private final AtomicReference<Instant> timestamp = new AtomicReference<>();
 	private final Addressing addressing;
